src/test/java/io/zipcoder/casino/PlayerTests/OU7PlayerTest.java
package io.zipcoder.casino.PlayerTests;

import io.zipcoder.casino.CasinoClasses.Membership;
import io.zipcoder.casino.Player.OU7Player;
import org.junit.Assert;
import org.junit.Test;

public class OU7PlayerTest {

    @Test
    //given
    public void OU7PlayerTest(Membership player) {
        Membership newMembership = new Membership("Fred Rogers", 123, 10000, 40);
        OU7Player newPlayer = new OU7Player(newMembership);

        //when
        String expectedName = "Fred Rogers";
        String actualName = newPlayer.getName();

        Integer actualAge = 40;
        Integer expectedAge = newPlayer.age;

        Integer actualIdNumber = 123;
        Integer expectedIdNumber = newPlayer.idNumber;
        Integer actualBalance = 10000;
        Integer expectedBalance = newPlayer.getBalance();

        //result
        Assert.assertEquals(actualName, expectedName);
        Assert.assertEquals(actualAge, expectedAge);
        Assert.assertEquals(actualIdNumber, expectedIdNumber);
        Assert.assertEquals(actualBalance, expectedBalance);


    }
}
